// src/types/filme.ts
export type Movie = {
    id: number;
    titulo: string;
    score: number;
    conta: number;
    image: string;
}

export type MoviePage = {
    content: Movie[];
    last: boolean;
    totalPages: number;
    totalElements: number;
    size: number;
    number: number;
    first: boolean;
    numberOfElements: number;
    empty: boolean;
}